// src/enum/keys.enum.ts
export const enum Keys {
  WINNING_COMBINATION = "Winning Combination(in any order)",
  WINNING_COMBINATION_EXACT = "Winning Combination(in exact order)",
  FIRST_PRIZE = "First Prize",
  JACKPOT_PRIZE = "Jackpot Prize",
  NUMBER_OF_WINNERS = "Number of Winner(s)",
  VISAYAS = "Visayas",
  MINDANAO = "Mindanao",
  // ? oldest date available
  OLD_WINNING_COMBINATION = "Winning Combination",
}